[Clara] Dear {REDACTED}
I do not know why it is that she wrote me.
I do not know how far this system reaches;
I only know that I must keep reaching, in the hopes that I find her;
to let her know she is not alone
to let her know she must also be a character
in need of forgiveness.

- Clara.